Update how to write patches
The "Vim development" page doesn't explain the detail now. Don't mention
it at the top of the page. (A link to the page is still kept at the
bottom of the page.)

Change Make_dos.mak to Make_mvc.mak to follow the latest source code.

Mention about Make_ming.mak.